Add files via upload
This branch has the desired behaviour of returning a new Hadith every 24hrs. For development purposes this duration has been changed to one minute but can easily be tweaked. The only issue with this version is that it spins up a new goroutine(with an infinite loop inside it) every time a get request is made which leaks memory.
It also lacks configuration for CORS.